+++ src/components/App.js
@@ -1,10 +1,10 @@
-import "./styles.css";
+import "../styles.css";
 import { Route, Switch } from "react-router-dom";
 import PlayVideo from "./PlayVideo";
 import Home from "./Home";
 import Account from "./Account";
 import Login from "./Login";
-import LikedVideos from "./components/LikedVideos";
+import LikedVideos from "../components/LikedVideos";
 import PrivateRoute from "./PrivateRoute";
 
 export default function App() {
