+++ src/index.js
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import { VideoProvider } from "./context/videoProvider";
-import App from "./App";
+import App from "./components/App";
 import { UrlProvider } from "./context/useVideoPlaylistId";
 import { AuthProvider } from "./context/AuthProvider";
 const rootElement = document.getElementById("root");
